+++ dto/src/main/kotlin/personDtos.kt
@@ -1,5 +1,5 @@
 package com.example.dto
 
-class UpdatePersonDto(val name: String?, val weight: Double?)
-class CreatePersonDto(val name: String, val weight: Double)
-class PersonDto(val id: Int, val name: String, val weight: Double)
+class UpdatePersonDto(val name: String?, val weight: Float?)
+class CreatePersonDto(val name: String, val weight: Float)
+class PersonDto(val id: Int, val name: String, val weight: Float)
